+++ src/pets/pets.service.ts
@@ -2,7 +2,6 @@ import { Injectable, NotFoundException } from '@nestjs/common';
 import { PetsRepository } from './pets.repository';
 import { CreatePetDto } from './dto/create-pet.dto';
 import { UpdatePetDto } from './dto/update-pet.dto';
-import { Prisma } from '@prisma/client';
 
 @Injectable()
 export class PetsService {
@@ -13,11 +12,18 @@ export class PetsService {
   }
 
   async findAll() {
-    return this.petsRepository.findAll();
+    const allPets = await this.petsRepository.findAll();
+
+    if (!allPets) {
+      throw new NotFoundException('No pets found');
+    }
+
+    return allPets;
   }
 
   async findOne(id: number) {
     const pet = await this.petsRepository.findOne(id);
+
     if (!pet) {
       throw new NotFoundException(`Pet with ID ${id} not found`);
     }
@@ -25,13 +31,22 @@ export class PetsService {
   }
 
   async update(id: number, updatePetDto: UpdatePetDto) {
-    await this.findOne(id);
+    const selectedPet = await this.findOne(id);
+
+    if (!selectedPet) {
+      throw new NotFoundException(`Pet with ID ${id} not found`);
+    }
 
     return this.petsRepository.update(id, updatePetDto);
   }
 
   async remove(id: number) {
-    await this.findOne(id);
+    const pet = await this.findOne(id);
+
+    if (!pet) {
+      throw new NotFoundException(`Pet with ID ${id} not found`);
+    }
+
     return this.petsRepository.remove(id);
   }
 }
